mea/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
faiss-cpu==1.8.0
transformers==4.41.0
torch==2.3.0
torchvision==0.18.0
anthropic==0.28.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
streamlit==1.35.0
requests==2.32.2
tqdm==4.66.4
scikit-learn==1.5.0
